==> analyses/molecular-subtyping-EPN/03-summary.Rmd <==
---
title: "Molecularly Subtyping EPN Tumors"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Komal S. Rathi (adapted from python notebook by Teja Koganti), Ryan Corbett
date: 2022
---

## Usage

This notebook is intended to be run via the command line from the top directory
of the repository as follows:

```
Rscript -e "rmarkdown::render('analyses/molecular-subtyping-EPN/03-summary.Rmd', clean = TRUE)"
```

## Set Up

### Libraries and functions

```{r}
suppressPackageStartupMessages({
  library(tidyverse)
})
```

### Directories

```{r}
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")
```

### Read in relevant files

```{r message=FALSE}
# Read in histologies file
path_dx_list <- jsonlite::fromJSON(file.path(root_dir, "analyses", "molecular-subtyping-EPN", "epn-subset", "EPN_subtyping_path_dx_strings.json"))
epn_samples <- read_tsv(file.path(data_dir, "histologies-base.tsv"), guess_max = 100000) %>%
  filter(pathology_diagnosis %in% path_dx_list$exact_path_dx) 

# Reading in table from results folder without the molecular_subtype column
EPN_final <- read_tsv("results/EPN_all_data.tsv")
```

### Spinal prioritization

We will first prioritize samples from spinal cord to be assigned to "EPN, SP-MYCN" (if MYCN amplification also observed) as described in [reference]. These samples will not be considered for further molecular_subtypeing. 
*"EPN, SP" molecular subtype is based on the methylation subtype, not by disease_group. The code was changed before. Investigation is needed.

```{r}
EPN_final <- EPN_final %>%
  mutate(molecular_subtype = case_when((disease_group == "spinal" & 
                                      grepl("amplification", consensus_focal_CN_MYCN)) | molecular_subtype_methyl == "EPN, SP-MYCN" ~ "EPN, SP-MYCN", 
                                      TRUE ~ NA_character_))
sp_mycn_samples_assigned <- EPN_final %>%
  filter(molecular_subtype == "EPN, SP-MYCN") %>%
  pull(match_id) %>%
  unique()
print(paste("Number of samples assigned for SP-MYCN:", length(sp_mycn_samples_assigned)))
```

### Fusion prioritization

We will next prioritize ZFTA (formaerly C11orf95) and YAP1 fusions based on evidence for these molecular_subtypes from Pajtler. et al. Fig.6

### ZFTA gene fusions

1. Here we are prioritizing ZFTA (formerly C11orf95) fusions, if any of the ZFTA fusions are present for a sample, it is assigned "EPN, ST ZFTA"
2. These samples will not be considered for further molecular_subtypeing

```{r}
EPN_final <- EPN_final %>%
  mutate(molecular_subtype = case_when(`ZFTA--RELA` == 1 | `ZFTA--MAML2` == 1 ~ "EPN, ST ZFTA",
                                      TRUE ~ molecular_subtype),
         # if any NA mol subtype, then add methyl subtype
         molecular_subtype = case_when(is.na(molecular_subtype) & molecular_subtype_methyl == "EPN, ST ZFTA" ~ "EPN, ST ZFTA",
                                              TRUE ~ molecular_subtype))
zfta_samples_assigned <- EPN_final %>%
  filter(molecular_subtype == "EPN, ST ZFTA") %>%
  pull(match_id) %>%
  unique()
print(paste("Number of samples assigned for ZFTA:", length(zfta_samples_assigned)))
```


### YAP1 gene fusions

1. Here we are prioritizing YAP1 fusions, if any of the YAP1 fusions are present for a sample that has not been molecular_subtypeed, it is assigned "EPN, ST YAP1"
2. These samples will not be considered for further molecular_subtypeing

```{r}
EPN_final <- EPN_final %>%
  mutate(molecular_subtype = case_when(`YAP1--MAMLD1` == 1 | `YAP1--MAML2` == 1 | `YAP1--FAM118B` == 1 ~ "EPN, ST YAP1",
                                       TRUE ~ molecular_subtype),
         # if any NA mol subtype, then add methyl subtype
         molecular_subtype = case_when(is.na(molecular_subtype) & molecular_subtype_methyl == "EPN, ST YAP1" ~ "EPN, ST YAP1",
                                              TRUE ~ molecular_subtype))

yap1_samples_assigned <- EPN_final %>%
  filter(molecular_subtype == "EPN, ST YAP1") %>%
  pull(match_id) %>%
  unique()
print(paste("Number of samples assigned for YAP1:", length(yap1_samples_assigned)))
```

### PT_EPN molecular_subtypes

1. Here we will assign the PT_EPN_A and PT_EPN_B molecular_subtypes to the sample that have not been molecular_subtyped before. 
2. Based on Pajtler et al. fig. 4, over expression of EZHIP (formerly CXorf67) and TKTL1 along with 1q gain is seen under PT_EPN_A molecular_subtype
3. Similarly GPBP1 and IFT46 shows over expression along with 6p and 6q loss in PT_EPN_B molecular_subtype


```{r}
EPN_final <- EPN_final %>%
  mutate(molecular_subtype = case_when(
    is.na(molecular_subtype) & `1q_gain` > 0 & TKTL1_expr_zscore > 3 ~  "EPN, PF A",
    is.na(molecular_subtype) & EZHIP_expr_zscore > 3 ~ "EPN, PF A",
    is.na(molecular_subtype) & CNS_region == "Posterior fossa" & 
      (EPN_final$`H3-3A_HGVSp_Short` == "p.K28M" | EPN_final$`H3-3B_HGVSp_Short` == "p.K28M" | H3C2_HGVSp_Short == "p.K28M" | 
         H3C3_HGVSp_Short == "p.K28M" | H3C14_HGVSp_Short == "p.K28M") ~ "EPN, PF A",
    is.na(molecular_subtype) & (`6q_loss` > 0 | `6p_loss` > 0) & 
      (GPBP1_expr_zscore > 3 | IFT46_expr_zscore > 3) ~ "EPN, PF B",
    is.na(molecular_subtype) & !is.na(molecular_subtype_methyl) ~ molecular_subtype_methyl,
    TRUE ~ molecular_subtype)
    )

epn_pfa_samples_assigned <- EPN_final %>%
  filter(molecular_subtype == "EPN, PF A") %>%
  pull(match_id) %>%
  unique()
print(paste("Number of samples assigned for PT_EPN_A:", length(epn_pfa_samples_assigned)))

epn_pfb_samples_assigned <- EPN_final %>%
  filter(molecular_subtype == "EPN, PF B") %>%
  pull(match_id) %>%
  unique()
print(paste("Number of samples assigned for PT_EPN_B:", length(epn_pfb_samples_assigned)))
```


# sort final table

```{r}
EPN_final_subtyped <- EPN_final %>% 
  mutate(molecular_subtype = case_when(is.na(molecular_subtype) ~ "EPN, To be classified",
                                             TRUE ~ molecular_subtype)) %>%
  arrange(Kids_First_Participant_ID, match_id)

# write out final file 
write_tsv(EPN_final_subtyped, "results/EPN_all_data_with_molecular_subtype.tsv")

# generate easy table for compile
EPN_subtypes_by_id <- EPN_final_subtyped %>%
  select(match_id, molecular_subtype, molecular_subtype_methyl) %>%
  unique()

# do we have 1 subtype per id?
length(unique(EPN_subtypes_by_id$match_id)) == nrow(EPN_subtypes_by_id)

EPN_subtypes_by_id <- EPN_subtypes_by_id %>%
  left_join(epn_samples[,c("Kids_First_Participant_ID", "Kids_First_Biospecimen_ID", "match_id", "sample_id")]) %>%
  arrange(Kids_First_Biospecimen_ID) %>%
  write_tsv("results/epn_molecular_subtypes.tsv")

# any missing? no
missing_subtypes <- setdiff(epn_samples$Kids_First_Biospecimen_ID, EPN_subtypes_by_id$Kids_First_Biospecimen_ID)

```

### Summary of final subtype counts

```{r}
EPN_subtypes_by_id %>%
  group_by(molecular_subtype) %>%
  summarise(n = n())

```
